requirements.txt
MacFSEvents==0.4
PyYAML==3.11
colorama==0.3.3
contextlib2==0.4.0
mock==1.0.1
nose==1.3.4
python-dateutil==2.4.0
python-termstyle==0.1.10
pytz==2014.10
requests==2.5.1
simplejson==3.6.5
six==1.9.0
sniffer==0.3.4
vcrpy==1.2.0
wrapt==1.10.2
wsgiref==0.1.2
